requirements.txt
langchain>=0.2.12
langgraph>=0.2.24
openai>=1.40.0
python-dotenv>=1.0.1
requests>=2.32.3
pydantic>=2.8.2
pandas>=2.2.2
tqdm>=4.66.4
chromadb>=0.5.3
lancedb>=0.7.0
gspread>=6.1.2
google-auth>=2.34.0
google-auth-oauthlib>=1.2.1
pyyaml>=6.0.2
google-generativeai>=0.8.2
streamlit>=1.38.0
